+++ SchoolTimetableApp/frmManageTeachers.vb
@@ -8,6 +8,9 @@
 
     Private Sub frmManageTeachers_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         sql.LoadComboBox("SELECT * FROM `class`;", cmbClass, "class")
+        If Not sql.result Then
+            Exit Sub
+        End If
         cmbClass.SelectedValue = cmbClass.Items(0).Row(0).ToString()
         refresh_teacher()
         refresh_data()
@@ -16,6 +19,7 @@
         ListView1.Columns.Add("To")
         ListView1.Columns.Add("Day", 100, HorizontalAlignment.Center)
 
+        cmbType.SelectedIndex = 0
     End Sub
 
     Private Sub cmbClass_SelectionChangeCommitted(sender As Object, e As EventArgs) Handles cmbClass.SelectionChangeCommitted
@@ -31,6 +35,9 @@
                             `class` c ON sc.`class_id` = c.`class_id`
                           WHERE
                             c.`class_id`='" & class_id & "';", cmbSubjects, "subject")
+        If Not sql.result Then
+            Exit Sub
+        End If
     End Sub
 
     Private Sub btnAdd_Click(sender As Object, e As EventArgs) Handles btnAdd.Click
@@ -49,6 +56,9 @@
             sql.readData("SELECT * FROM `teacher_class` WHERE `teacher_id`='" & cmbName.SelectedValue.ToString() & "' AND
                         `subject_id`='" & cmbSubjects.SelectedValue.ToString() & "' AND
                         `class_id` = '" & cmbClass.SelectedValue.ToString() & "';")
+            If Not sql.result Then
+                Exit Sub
+            End If
             If sql.rd.HasRows() Then
                 MsgBox(cmbSubjects.SelectedItem.row(1).ToString() & " has already been added for " & cmbClass.SelectedItem.row(1).ToString())
                 sql.CloseCon()
@@ -60,6 +70,9 @@
                 ParamValues.Add(cmbSubjects.SelectedValue.ToString())
                 ParamValues.Add(cmbClass.SelectedValue.ToString())
                 sql.InsertQuery("INSERT INTO `teacher_class`(`teacher_id`,`subject_id`,`class_id`) VALUES(@1,@2,@3);", ParamValues)
+                If Not sql.result Then
+                    Exit Sub
+                End If
                 refresh_data()
 
             End If
@@ -82,6 +95,9 @@
                                     LEFT JOIN
                                 `subjects` s ON s.`subject_id` = tc.`subject_id`
                                 ORDER BY c.`name`;", dgvTeachersList, "Teachers")
+        If Not sql.result Then
+            Exit Sub
+        End If
 
         dgvTeachersList.Columns(0).Visible = False
         dgvTeachersList.Columns(1).Width = 200
@@ -96,6 +112,9 @@
         Dim teacher_id As Long = 0
 
         sql.readData("SELECT * FROM `teachers` WHERE `name`='" & txtName.Text & "';")
+        If Not sql.result Then
+            Exit Sub
+        End If
         If sql.rd.HasRows() Then
             MsgBox("Teacher already exists in  the database")
             sql.CloseCon()
@@ -105,10 +124,20 @@
             ParamValues.Add(txtName.Text)
             ParamValues.Add(cmbType.SelectedItem.ToString())
             teacher_id = Long.Parse(sql.InsertQuery("INSERT INTO `teachers`(`name`,`type`) VALUES(@1,@2);", ParamValues))
+            If Not sql.result Then
+                Exit Sub
+            End If
 
-            If teacher_id <> 0 Then
+            If teacher_id <> 0 And cmbType.SelectedIndex <> 0 Then
+                ParamValues.Clear()
+                ParamValues.Add(Trim(txtFrom.Text))
+                ParamValues.Add(Trim(txtTo.Text))
+                ParamValues.Add(cmbDay.SelectedItem.ToString())
                 For Each item As ListViewItem In ListView1.Items
-                    sql.ExecuteQuery("UPDATE `teachers` SET `from` = '" & item.SubItems(0).ToString() & "', `to` = '" & item.SubItems(1).ToString() & "', `day` = '" & item.SubItems(2).ToString() & "' WHERE `teacher_id`='" & teacher_id & "'; ")
+                    sql.InsertQuery("INSERT INTO `teacher_available_time`(`from`, `to`,`day`) VALUES (@1,@2,@3); ", ParamValues)
+                    If Not sql.result Then
+                        Exit Sub
+                    End If
                 Next
             End If
             MsgBox("Teacher successfully added",, "Manage Teachers")
@@ -119,6 +148,9 @@
 
     Private Sub refresh_teacher()
         sql.LoadComboBox("SELECT * FROM `teachers` ORDER BY `name`;", cmbName, "teacher")
+        If Not sql.result Then
+            Exit Sub
+        End If
     End Sub
 
     Private Sub btnDelete_Click(sender As Object, e As EventArgs) Handles btnDelete.Click
@@ -126,9 +158,16 @@
 
         If MsgBox("Are you sure you want to delete teacher" & vbCrLf & "This will remove the teacher's name from the database and unassign all subjects assigned to the teacher", MsgBoxStyle.Question & MsgBoxStyle.YesNo, "Manage Teachers") = MsgBoxResult.Yes Then
             sql.ExecuteQuery("DELETE FROM `teachers` WHERE `teacher_id`='" & id & "'")
+            If Not sql.result Then
+                Exit Sub
+            End If
             sql.ExecuteQuery("DELETE FROM `teacher_class` WHERE `teacher_id`='" & id & "'")
+            If Not sql.result Then
+                Exit Sub
+            End If
             MsgBox("Teacher successfully deleted")
             refresh_data()
+            refresh_teacher()
         End If
     End Sub
 
@@ -145,6 +184,9 @@
             Dim teacher_name As String = dgvTeachersList.Rows(e.RowIndex).Cells(e.ColumnIndex).Value.ToString()
             Dim teacher_id As Integer = CInt(dgvTeachersList.Rows(e.RowIndex).Cells(0).Value.ToString())
             sql.ExecuteQuery("UPDATE `teachers` SET `name`='" & teacher_name & "' WHERE `teacher_id`='" & teacher_id & "';")
+            If Not sql.result Then
+                Exit Sub
+            End If
             MsgBox("Update successful")
             refresh_data()
         End If
@@ -157,6 +199,9 @@
         Dim teacher_class As String = dgvTeachersList.CurrentRow.Cells(2).Value.ToString()
 
         sql.readData("SELECT `subject_id` FROM `subjects` WHERE `name`='" & subject & "';")
+        If Not sql.result Then
+            Exit Sub
+        End If
         If sql.rd.HasRows() Then
             sql.rd.Read()
             Dim subject_id As Integer = sql.rd.GetInt32("subject_id")
@@ -164,6 +209,9 @@
 
             If MsgBox("Do you really want to unassign all subjects from " & teacher, MsgBoxStyle.Question & MsgBoxStyle.YesNo, "Manage Teachers") = MsgBoxResult.Yes Then
                 sql.ExecuteQuery("DELETE FROM `teacher_class` WHERE `teacher_id`='" & id & "' AND `subject_id`='" & subject_id & "';")
+                If Not sql.result Then
+                    Exit Sub
+                End If
                 MsgBox(subject & " successfully removed from Teacher across all classes")
                 refresh_data()
             End If
@@ -180,6 +228,9 @@
 
 
         sql.readData("SELECT `class_id` FROM `class` WHERE `name`='" & teacher_class & "';")
+        If Not sql.result Then
+            Exit Sub
+        End If
         If sql.rd.HasRows() Then
             sql.rd.Read()
             Dim class_id As Integer = sql.rd.GetInt32("class_id")
@@ -187,6 +238,9 @@
 
             If MsgBox("Are you sure you want to unassign all subjects from " & teacher & " for " & teacher_class, MsgBoxStyle.Question & MsgBoxStyle.YesNo, "Manage Teachers") = MsgBoxResult.Yes Then
                 sql.ExecuteQuery("DELETE FROM `teacher_class` WHERE `teacher_id`='" & id & "' AND `class_id`='" & class_id & "';")
+                If Not sql.result Then
+                    Exit Sub
+                End If
                 MsgBox("Subjects successfully removed from " & teacher & " for " & teacher_class)
                 refresh_data()
             End If
@@ -209,6 +263,9 @@
         If MsgBox("Are you sure you want to truncate this table?" & vbCrLf & "This will delete all records", MsgBoxStyle.Question & MsgBoxStyle.YesNo) = MsgBoxResult.Yes Then
 
             sql.ExecuteQuery("TRUNCATE TABLE `teacher_class`;")
+            If Not sql.result Then
+                Exit Sub
+            End If
             MsgBox("Truncate operation successful")
             refresh_data()
 
@@ -216,10 +273,17 @@
     End Sub
 
     Private Sub btnTruncateTeachers_Click(sender As Object, e As EventArgs) Handles btnTruncateTeachers.Click
-
-        sql.ExecuteQuery("TRUNCATE TABLE `teachers`;")
-        MsgBox("Teacher List successfully truncated")
-        refresh_teacher()
+        If MsgBox("Are you sure you want to truncate this table?" & vbCrLf & "This will delete all records of teachers", MsgBoxStyle.Question & MsgBoxStyle.YesNo) = MsgBoxResult.Yes Then
+            sql.ExecuteQuery("TRUNCATE TABLE `teachers`;")
+            If Not sql.result Then
+                Exit Sub
+            End If
+            MsgBox("Teacher List successfully truncated")
+            If Not sql.result Then
+                Exit Sub
+            End If
+            refresh_teacher()
+        End If
     End Sub
 
     Private Sub dgvTeachersList_CellContentClick(sender As Object, e As DataGridViewCellEventArgs) Handles dgvTeachersList.CellContentClick
@@ -236,6 +300,9 @@
 
 
         sql.readData("SELECT `class_id` FROM `class` WHERE `name`='" & teacher_class & "';")
+        If Not sql.result Then
+            Exit Sub
+        End If
         If sql.rd.HasRows() Then
             sql.rd.Read()
             class_id = sql.rd.GetInt32("class_id")
@@ -243,6 +310,9 @@
         End If
 
         sql.readData("SELECT `subject_id` FROM `subjects` WHERE `name`='" & subject & "';")
+        If Not sql.result Then
+            Exit Sub
+        End If
         If sql.rd.HasRows() Then
             sql.rd.Read()
             subject_id = sql.rd.GetInt32("subject_id")
@@ -252,6 +322,9 @@
         If class_id <> 0 And subject_id <> 0 Then
             If MsgBox("Are you sure you want to unassign " & subject & " from " & teacher & " for " & teacher_class, MsgBoxStyle.Question & MsgBoxStyle.YesNo, "Manage Teachers") = MsgBoxResult.Yes Then
                 sql.ExecuteQuery("DELETE FROM `teacher_class` WHERE `teacher_id`='" & id & "' AND `class_id`='" & class_id & "' AND `subject_id`='" & subject_id & "';")
+                If Not sql.result Then
+                    Exit Sub
+                End If
                 MsgBox("Subject successfully removed ")
                 refresh_data()
             End If
@@ -268,6 +341,38 @@
     End Sub
 
     Private Sub btnDeleteTime_Click(sender As Object, e As EventArgs) Handles btnDeleteTime.Click
+        For index = 0 To ListView1.SelectedItems.Count - 1
+            ListView1.Items.RemoveAt(ListView1.SelectedIndices(0))
+        Next
+    End Sub
 
+    Private Sub cmbType_SelectedIndexChanged(sender As Object, e As EventArgs) Handles cmbType.SelectedIndexChanged
+        If cmbType.SelectedIndex <> 0 Then
+            gbAvailableTeacherTime.Visible = True
+        Else
+            gbAvailableTeacherTime.Visible = False
+        End If
+    End Sub
+
+    Private Sub btnUpdateTeacher_Click(sender As Object, e As EventArgs) Handles btnUpdateTeacher.Click
+        frmUpdateTeacherRecord.teacher_id = Integer.Parse(cmbName.SelectedValue.ToString())
+        frmUpdateTeacherRecord.Show()
+        Hide()
+    End Sub
+
+    Private Sub btnDeleteTeacherRecord_Click(sender As Object, e As EventArgs) Handles btnDeleteTeacherRecord.Click
+        If MsgBox("Are you sure you want to delete teacher" & vbCrLf & "This will remove the teacher's name from the database and unassign all subjects assigned to the teacher", MsgBoxStyle.Question & MsgBoxStyle.YesNo, "Manage Teachers") = MsgBoxResult.Yes Then
+            sql.ExecuteQuery("DELETE FROM `teachers` WHERE `teacher_id`='" & cmbName.SelectedValue.ToString() & "'")
+            If Not sql.result Then
+                Exit Sub
+            End If
+            sql.ExecuteQuery("DELETE FROM `teacher_class` WHERE `teacher_id`='" & cmbName.SelectedValue.ToString() & "'")
+            If Not sql.result Then
+                Exit Sub
+            End If
+            MsgBox("Teacher successfully deleted")
+            refresh_teacher()
+            refresh_data()
+        End If
     End Sub
 End Class
